+++ src/main/kotlin/de/livepoll/api/entity/dto/MultipleChoiceItemDtoIn.kt
@@ -1,7 +1,7 @@
 package de.livepoll.api.entity.dto
 
 data class MultipleChoiceItemDtoIn(
-        val pollId: Int,
+        val pollId: Long,
         val question: String,
         val position: Int,
         val allowMultipleAnswers: Boolean,
